- Enhance the Employee model for better currency precision and establish
a relationship with timesheets.
- Introduce product and purchase models with corresponding relationships
and seeders, improving data management and integrity.
- Refresh database seeders for realistic data entries.

+++ source_code/database/migrations/2025_10_08_222413_create_employees_table.php
@@ -11,7 +11,7 @@ return new class extends Migration {
 			$table->unsignedBigInteger('id')->primary(); // This is also the foreign key to users table
 			$table->foreign('id')->references('id')->on('users'); // Foreign key constraint
 			$table->string('phone')->unique();
-			$table->double('hourly_wage');
+			$table->decimal('hourly_wage', 10, 2);
 			$table->string('status');
 			$table->string('payment_frequency');
 			$table->timestamps();
